requirements.txt
absl-py==2.1.0
numpy==1.26.4
pydantic==2.10.6
matplotlib==3.9.2
parallelbar==2.5
ray[default]==2.46.0
scipy==1.15.3
tqdm==4.66.5
pandas==2.1.4
pyarrow==20.0.0
Cython==3.0.11
